fixed UI issues and table arrangments

+++ src/store/modules/dinein.js
@@ -113,7 +113,7 @@ const actions = {
       table.related_orders_ids.forEach(order_Id => {
         let od = response.data.page_lookups.orders._id[order_Id]
         order.push(od)
-        balanceDue += parseInt(od.balance_due)
+        balanceDue += parseFloat(od.balance_due)
         currency = od.currency
       })
       orderDetails.push({
